moved etcdclient to utils
